modified CMakeLists
