+++ .clang-tidy
@@ -1,0 +1,18 @@
+Checks: >
+  *,
+  bugprone-*,
+  performance-*,
+  readability-*,
+  modernize-*,
+  cppcoreguidelines-*,
+  clang-analyzer-*,
+  -modernize-use-trailing-return-type,
+  -modernize-use-equals-default,
+  -misc-const-correctness
+
+
+HeaderFilterRegex: '^(?!.*(_deps|external)).*'
+
+WarningsAsErrors: ''
+
+FormatStyle: 'file'
