+++ .gitignore
@@ -1,0 +1,4 @@
+#Directories
+build/
+.vscode/
+*.html
