--- include/PrintRandomNumbers.hpp
@@ -1,27 +1,0 @@
-#ifndef PRINTRANDOMNUMBERS_HPP
-#define PRINTRANDOMNUMBERS_HPP
-
-#include <iostream>
-#include <vector>
-#include "RandomGenerator.hpp"
-
-
-class PrintRandomNumbers 
-{
-    public:
-    PrintRandomNumbers(RandomGeneratorAbstract& random);
-    ~PrintRandomNumbers() = default;
-
-    void printChoosenNumbers(unsigned int columns);
-    void readRandomNumbers();
-    void saveChoosenNumbers();
-
-    private:
-    std::vector<int> randomData;
-    RandomGeneratorAbstract& randomGenerator;
-    int minVal;
-    int maxVal;
-    int nrSamples;
-};
-
-#endif //PRINTRANDOMNUMBERS_HPP
